fix(stdint): change unsigned types for stdint types when allignment is key to prevent bad compilations (eg. unsigned int -> uint32_t ensures 32 bits)

+++ os/include/mem/e820.h
@@ -6,16 +6,18 @@
 #define E820_TYPE_USABLE    1
 #define E820_TYPE_RESERVED  2
 
+#include <stdint.h>
+
 // Structure representing a single entry returned by BIOS INT 0x15 E820
 typedef struct {
-    unsigned long long base_addr;
-    unsigned long long length;
-    unsigned int type;
+    uint64_t base_addr;
+    uint64_t length;
+    uint32_t type;
 } __attribute__((packed)) e820_entry_t;
 
 // Memory map info container populated in bootloader
 typedef struct {
-    unsigned int count;
+    uint32_t count;
     e820_entry_t entries[32]; // Max entries stored at 0x8004
 } __attribute__((packed)) e820_map_t;
